Done with first 2 exercises: fixed bug with roots

+++ HW03/QuadraticEquation.h
@@ -4,20 +4,22 @@
 
 using namespace std;
 
+//creates QuadraticEquation class
 class QuadraticEquation
 {
+//declares the values of a, b, and c as private
 private:
-	int a, b, c;
+	double a, b, c;
+//functions are made public
 public:
-	int geta();
-	int getb();
-	int getc();
-	QuadraticEquation(int a1, int b1, int c1);
-	double getDiscriminant(int a, int b, int c);
-	double getRoot1(int a, int b, double discriminant);
-	double getRoot2(int a, int b, double discriminant);
-
-
-	
-
+	//declares get functions for the values of a, b, and c
+	double geta();
+	double getb();
+	double getc();
+	//function for QuadraticEquation using new user input values
+	QuadraticEquation(double a1, double b1, double c1);
+	//declares get functions for the discriminant and the two roots
+	double getDiscriminant(double a, double b, double c);
+	double getRoot1(double a, double b, double discriminant);
+	double getRoot2(double a, double b, double discriminant);
 };
